refactor(play): add get-increment
